+++ packages/@baldeweg/ui/src/components/__tests__/BHorizontalList.spec.js
@@ -1,9 +1,0 @@
-import { describe, it, expect } from 'vitest'
-
-import BHorizontalList from '../BHorizontalList/BHorizontalList.vue'
-
-describe('BHorizontalList', () => {
-  it('shows BHorizontalList', () => {
-    expect(BHorizontalList).toBeTruthy()
-  })
-})