+++ packages/@baldeweg/ui/src/components/__tests__/BHorizontalListItem.spec.js
@@ -1,9 +1,0 @@
-import { describe, it, expect } from 'vitest'
-
-import BHorizontalListItem from '../BHorizontalList/BHorizontalListItem.vue'
-
-describe('BHorizontalListItem', () => {
-  it('shows BHorizontalListItem', () => {
-    expect(BHorizontalListItem).toBeTruthy()
-  })
-})