+++ packages/@baldeweg/ui/src/components/__tests__/BSlider.spec.js
@@ -1,0 +1,9 @@
+import { describe, it, expect } from 'vitest'
+
+import BSlider from '../BSlider/BSlider.vue'
+
+describe('BSlider', () => {
+  it('shows BSlider', () => {
+    expect(BSlider).toBeTruthy()
+  })
+})