+++ packages/@baldeweg/ui/src/components/__tests__/BSliderItem.spec.js
@@ -1,0 +1,9 @@
+import { describe, it, expect } from 'vitest'
+
+import BSliderItem from '../BSlider/BSliderItem.vue'
+
+describe('BSliderItem', () => {
+  it('shows BSliderItem', () => {
+    expect(BSliderItem).toBeTruthy()
+  })
+})