+++ packages/@baldeweg/ui/src/views/BSliderView.vue
@@ -1,80 +1,80 @@
 <script setup>
 import { useTitle } from './../composables/useTitle.js'
 
-useTitle({ title: 'BHorizontalList - @baldeweg/ui' })
+useTitle({ title: 'BSlider - @baldeweg/ui' })
 </script>
 
 <template>
   <article>
-    <BHorizontalList>
-      <BHorizontalListItem size="l">
+    <BSlider>
+      <BSliderItem size="l">
         <BDivider />
         <p>Item 1</p>
         <p>Description</p>
-      </BHorizontalListItem>
-      <BHorizontalListItem size="s">
+      </BSliderItem>
+      <BSliderItem size="s">
         <BDivider />
         <p>Item 2</p>
         <p>Description</p>
-      </BHorizontalListItem>
-      <BHorizontalListItem size="m">
+      </BSliderItem>
+      <BSliderItem size="m">
         <BDivider />
         <p>Item 3</p>
         <p>Description</p>
-      </BHorizontalListItem>
-      <BHorizontalListItem size="xs">
+      </BSliderItem>
+      <BSliderItem size="xs">
         <BDivider />
         <p>Item 4</p>
         <p>Description</p>
-      </BHorizontalListItem>
-      <BHorizontalListItem size="xl">
+      </BSliderItem>
+      <BSliderItem size="xl">
         <BDivider />
         <p>Item 5</p>
         <p>Description</p>
-      </BHorizontalListItem>
-      <BHorizontalListItem>
+      </BSliderItem>
+      <BSliderItem>
         <BDivider />
         <p>Item 6</p>
         <p>Description</p>
-      </BHorizontalListItem>
-      <BHorizontalListItem>
+      </BSliderItem>
+      <BSliderItem>
         <BDivider />
         <p>Item 7</p>
         <p>Description</p>
-      </BHorizontalListItem>
-      <BHorizontalListItem>
+      </BSliderItem>
+      <BSliderItem>
         <BDivider />
         <p>Item 8</p>
         <p>Description</p>
-      </BHorizontalListItem>
-      <BHorizontalListItem>
+      </BSliderItem>
+      <BSliderItem>
         <BDivider />
         <p>Item 9</p>
         <p>Description</p>
-      </BHorizontalListItem>
-      <BHorizontalListItem>
+      </BSliderItem>
+      <BSliderItem>
         <BDivider />
         <p>Item 10</p>
         <p>Description</p>
-      </BHorizontalListItem>
-    </BHorizontalList>
+      </BSliderItem>
+    </BSlider>
 
-    <BHorizontalList>
-      <BHorizontalListItem size="m">
+    <BSlider>
+      <BSliderItem size="m">
         <BDivider />
         <p>Item 1</p>
         <p>Description</p>
-      </BHorizontalListItem>
-      <BHorizontalListItem size="m">
+      </BSliderItem>
+      <BSliderItem size="m">
         <BDivider />
         <p>Item Item Item Item Item Item Item Item Item Item</p>
         <p>Description</p>
-      </BHorizontalListItem>
-      <BHorizontalListItem size="m">
+      </BSliderItem>
+      <BSliderItem size="m">
         <BDivider />
         <p>Item 3</p>
         <p>Description</p>
-      </BHorizontalListItem>
-    </BHorizontalList>
+      </BSliderItem>
+    </BSlider>
   </article>
 </template>