+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## v0.2.0
 
 - [ui] BModal can be closed with esc key
+- [ui] BSlider replaces BHorizontalList